recipepage basics

+++ code/Reciply/Reciply/Models/Ingredient.cs
@@ -8,7 +8,7 @@ namespace Reciply.Models
     {
         public int Id { get; set; }
         public string Item { get; set; } //Name des zu einkaufenden Gegenstandes (z.B. Mehl)
-        public int Amount { get; set; } //Wie viel davon einzukaufen ist (z.B. 1{kg})
+        public double Amount { get; set; } //Wie viel davon einzukaufen ist (z.B. 1{kg})
         public UnitOfMeasurement UnitOfMeasurement { get; set; } //Die Einheit von amount (z.B. kg,ml,l,...)
         public bool IsInShoppingBasket { get; set; } //True = Wenn der Gegenstand in der Einkaufsliste abgehakt ist (im Einkaufskorb), False = Wenn nicht
     }
